Move all discounts to a package.

+++ src/main/java/org/example/Product.java
@@ -1,3 +1,3 @@
 package org.example;
 
-public record Product(String name, double price, double quantity) {}
+public record Product(String name, double price, int quantity) {}
